app list improvements
