+++ src/components/app/list/AppListView.tsx
@@ -10,7 +10,6 @@ import { Empty } from './emptyView/Empty';
 import { URLS } from '../../../config';
 import { App, AppListState, OrderBy, SortBy } from './types';
 import { ReactComponent as Edit } from '../../../assets/icons/ic-settings.svg';
-import { ReactComponent as Info } from '../../../assets/icons/ic-info-outline.svg';
 import { ReactComponent as Search } from '../../../assets/icons/ic-search.svg';
 import { ReactComponent as Clear } from '../../../assets/icons/ic-error.svg';
 import { TriggerInfoModal } from './TriggerInfo';
@@ -245,27 +244,20 @@ export class AppListView extends Component<AppListViewProps>{
         else if (this.props.view === AppListViewType.EMPTY) {
             return <React.Fragment>
                 {this.renderPageHeader()}
-                {this.props.isDockerRegistryEmpty && <p className="m-0 pt-10 pb-10 pl-24 pr-24 cn-9 bcb-1 fw-5 fs-13">
-                    <Info className="icon-dim-20 fcb-5 vertical-align-bottom mr-8" />
-                    Docker registry configuration is required to setup a new application.
-                    &nbsp;<Link to={URLS.GLOBAL_CONFIG_DOCKER} className="cb-5 anchor">Click here to configure docker registry.</Link>
-                </p>}
                 {this.renderRouter()}
-                <Empty view={this.props.view}
-                    title={"No Applications here"}
-                    message={"You don't have any apps as of now."}
-                    buttonLabel={"Add new app"}
-                    clickHandler={this.openCreateModal} />
+                <AppCheckListModal history={this.props.history}
+                    location={this.props.location}
+                    match={this.props.match}
+                    appChecklist={this.props.appChecklist}
+                    chartChecklist={this.props.chartChecklist}
+                    appStageCompleted={this.props.appStageCompleted}
+                    chartStageCompleted={this.props.chartStageCompleted}
+                />
             </React.Fragment>
         }
         else if (this.props.view === AppListViewType.NO_RESULT) {
             return <React.Fragment>
                 {this.renderPageHeader()}
-                {this.props.isDockerRegistryEmpty && <p className="m-0 pt-10 pb-10 pl-24 pr-24 cn-9 bcb-1 fw-5 fs-13">
-                    <Info className="icon-dim-20 fcb-5 vertical-align-bottom mr-8" />
-                    Docker registry configuration is required to setup a new application.
-                    &nbsp;<Link to={URLS.GLOBAL_CONFIG_DOCKER} className="cb-5 anchor">Click here to configure docker registry.</Link>
-                </p>}
                 {this.renderSavedFilters()}
                 {this.renderRouter()}
                 <Empty view={this.props.view}
@@ -284,17 +276,6 @@ export class AppListView extends Component<AppListViewProps>{
         else {
             return <React.Fragment>
                 {this.renderPageHeader()}
-                {this.props.isDockerRegistryEmpty && <p className="m-0 pt-10 pb-10 pl-24 pr-24 cn-9 bcb-1 fw-5 fs-13">
-                    <Info className="icon-dim-20 fcb-5 vertical-align-bottom mr-8" />
-                    Docker registry configuration is required to setup a new application.
-                    &nbsp;<Link to={URLS.GLOBAL_CONFIG_DOCKER} className="cb-5 anchor">Click here to configure docker registry.</Link>
-                </p>}
-                <AppCheckListModal history={this.props.history} location={this.props.location} match={this.props.match}
-                    appChecklist={this.props.appChecklist}
-                    chartChecklist={this.props.chartChecklist}
-                    appStageCompleted={this.props.appStageCompleted}
-                    chartStageCompleted={this.props.chartStageCompleted}
-                />
                 {this.renderRouter()}
                 {this.renderSavedFilters()}
                 {this.renderAppList()}
